+++ app/src/org/commcare/heartbeat/HeartbeatRequester.java
@@ -36,8 +36,6 @@ import java.util.HashMap;
 public class HeartbeatRequester {
 
     private static final String TAG = HeartbeatRequester.class.getSimpleName();
-    private static final String TEST_RESPONSE =
-            "{\"app_id\":\"73d5f08b9d55fe48602906a89672c214\",\"latest_apk_version\":{\"value\":\"2.36.1\"},\"latest_ccz_version\":{\"value\":\"75\", \"force_by_date\":\"2017-05-01\"}}";
 
     private static final String APP_ID = "app_id";
     private static final String QUARANTINED_FORMS_PARAM = "num_quarantined_forms";
@@ -95,25 +93,6 @@ public class HeartbeatRequester {
                     "Received error response from heartbeat request: " + responseCode);
         }
     };
-
-    protected static void parseTestHeartbeatResponse() {
-        System.out.println("NOTE: Testing heartbeat response processing");
-        try {
-            parseHeartbeatResponse(new JSONObject(TEST_RESPONSE));
-        } catch (JSONException e) {
-            System.out.println("Test response was not properly formed JSON");
-        }
-    }
-
-    protected static void simulateRequestGettingStuck() {
-        System.out.println("Before sleeping");
-        try {
-            Thread.sleep(5*1000);
-        } catch (InterruptedException e) {
-            System.out.println("TEST ERROR: sleep was interrupted");
-        }
-        System.out.println("After sleeping");
-    }
 
     protected void requestHeartbeat() {
         String urlString = CommCareApplication.instance().getCurrentApp().getAppPreferences()
